Add files via upload
Java Program - 2/9/2021

Write a program that plays the popular scissor-rock-paper game (A scissor can cut a paper
a rock can knock a scissor, and a paper can wrap a rock.  The program randomly
generates a number 0,1,2 
0 = scissor
1 = rock
2 = paper

as the computers choice.  The program prompts the user to enter a number 0,1, or 2 and displays
messages indicating whether the user or the computer wins, loses or draws their choice.  
The program will the user continuously play until  either the user or the computer wins a
total three times.