Add files via upload
Java Assignment - 3/12/2021

(The Course class)

Revise the Course class as follows:

(DON’T use ArrayList)

The array size is fixed in Listing 10.6. Improve it to automatically increase the array size by creating a new larger array and copying the contents of the current array to it. ( default capacity = 16, increased the size of array to the double if necessary)
Implement the dropStudent method. If the student to be dropped can’t be found, do nothing.
Add the instructor’s information into the class.
Add a method, clear(), removes all students from the course.
Add a method, print(), displays course information with the Course Name, Instructor's name, the number of students enrolled, and the list of students' names.
Write a test program (Assignment5.java) that creates a course with default capacity, adds twenty students, removes three, adds another two, then displays the course info. Call clear() and display the course info again.