Add files via upload
Java Assignment - COMSC 255

This assignment was locked Mar 9 at 11:59pm.
(The Account class)

Design a class named Account that contains:

■ A private int data field named id for the account (default 0).

■ A private double data field named balance for the account (default 0).

■ A private double data field named annualInterestRate that stores the current interest rate (default 0). Assume all accounts have the same interest rate.

■ A private Date data field named dateCreated that stores the date when the account was created.

■ A no-arg constructor that creates a default account.

■ A constructor that creates an account with the specified id and initial balance.

■ The accessor and mutator methods for id, balance, and annualInterestRate.

■ The accessor method for dateCreated.

■ A method named getMonthlyInterestRate() that returns the monthly interest rate.

■ A method named getMonthlyInterest() that returns the monthly interest.

■ A method named withdraw that withdraws a specified amount from the account.

■ A method named deposit that deposits a specified amount to the account.

(Hints: )

The method getMonthlyInterest() is to return monthly interest, not the interest rate. Monthly interest is balance * monthlyInterestRate. monthlyInterestRate is annualInterestRate / 12. Note that annualInterestRate is a percentage, e.g.,like 3%. You need to divide it by 100.)
The value of dateCreated should be assigned in the constructor
Apply the sample test driver (Assignment4.java  download) that set the annual interest rate of 3%, creates an Account object array of size 5 (the account IDs are 1000 to 5000, balance of $1,000 to 5000),  invokes the withdraw method to withdraw $500, and the deposit method to deposit $1000 through each object. It also displays the account information in a for-loop.
Refer to the sample test driver, implement and use the static  displayAccounts(Account[]) method to replace the for-loop. The static method will print the balance, the monthly interest, and the date when this account was created for each account by passing the Account array in a table format.